pyproject.toml
[tool.poetry]
name = "langgraph-study"
version = "0.1.0"
description = ""
authors = ["Will <yzj_000@163.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.9"
langgraph = ">=0.5.1,<0.6.0"
langchain-deepseek = "0.1.3"
python-dotenv = ">=1.1.0,<1.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
